SMILE-687-search-tag-filter-autocomplete-bug, fixed-but, added test, refactor tags-filter tests
