+++ cypress/integration/dashboard/resetSearch.spec.js
@@ -1,5 +1,5 @@
-import { SearchParameters } from '../../src/types/search.types';
-import { ResourceFeatureTypes } from '../../src/types/resource.types';
+import { SearchParameters } from '../../../src/types/search.types';
+import { ResourceFeatureTypes } from '../../../src/types/resource.types';
 
 context('Actions', () => {
   const search = 'bananas';
